Selesai

+++ home/homeadmin.php
@@ -24,37 +24,41 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
     <title>Portal Akademik UAS</title>
-
-    <style>
-        .contrainer{
-            display: flex;
-            justify-content: center;
-            align-items: center;
-        }
-    </style>
-    <script>
-        function newDoc(a){}
-    </script>
 </head>
 
 <body>
-    <nav class="navbar navbar-expand-lg text-bg-warning p-3">
-    <div class="container-fluid">
-        <a class="navbar-brand" href="/home/homeadmin.php">Universitas Alim Sukses</a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-        </ul>
-        <form class="d-flex" role="search" action="homeadmin.php" method="GET">
-            <input class="form-control me-2" type="search" name="search" placeholder="Pencarian berdasarkan Nama" aria-label="Search">
-            <button class="btn btn-outline-dark me-2" type="submit">Search</button>
-            <a href="../action/logout.php" class="btn btn-primary">Logout</a>
-        </form>
-        </div>
-    </div>
-    </nav>
+    <header>
+        <nav class="navbar navbar-expand-md navbar-dark bg-dark">
+          <div class="container-fluid">
+            <a class="navbar-brand" href="/home/homeadmin.php">Universitas Alim Sukses</a>
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarCollapse" aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
+              <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarCollapse">
+              <ul class="navbar-nav me-auto mb-2 mb-md-0">
+                <li class="nav-item">
+                  <!-- <a class="nav-link active" aria-current="page" href="#">Home</a> -->
+                </li>
+                <li class="nav-item">
+                  <!-- <a class="nav-link" href="#">Link</a> -->
+                </li>
+                <li class="nav-item">
+                  <!-- <a class="nav-link disabled">Disabled</a> -->
+                </li>
+              </ul>
+                <!-- Button trigger modal -->
+                <form class="d-flex" role="search" action="homeadmin.php" method="GET">
+                    <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search" name="search"> 
+                    <button class="btn btn-outline-success me-2" type="submit">Search</button>
+                    <a href="../action/logout.php" class="btn btn-outline-success">Logout</a>
+                </form>
+            </div>
+          </div>
+        </nav>
+    </header>
+
+
+
     <div class="container">
     <table class="table table-bordered">
             <tr>
@@ -75,7 +79,7 @@
                             echo "<td>" . $user_krs['tgl_lahir'] . "</td>";
                             ?>
 
-                            <td><a class='btn btn-primary' href="<?php echo '../action/edit.php?id='. $user_krs['id']; ?>">Edit </a> | <a class='btn btn-primary' onclick="newDoc('<?php echo $user_krs['id']; ?>')" >Delete</a></td></tr>
+                            <td><a class='btn btn-primary' href="<?php echo '../action/edit.php?id='. $user_krs['id']; ?>">Edit </a> | <a class='btn btn-primary' onclick="newDoc('<?php echo $user_krs['id']; ?>', '<?php echo $user_krs['nama']; ?>')" >Delete</a></td></tr>
                             <?php
                         }
                     }else{
@@ -89,9 +93,11 @@
         </table>
     </div>
     <script>
-        function newDoc(a){
-            alert("Apakah anda ingin menghapus barang?");
-            window.location.href="../action/delete.php?id=" + a ;  
+        function newDoc(a, nama){
+            if(confirm("Apakah anda ingin menghapus " + nama + "?")){
+                window.location.href="../action/delete.php?id=" + a ; 
+            }
+                 
         }
 
     </script>
